% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_count.R
\docType{data}
\name{covid19_count}
\alias{covid19_count}
\title{COVID-19 case numbers in three countries}
\format{
In total there are 1395 observations across 7 variables.
\itemize{
\item \strong{Type}: coronavirus case type being confirmed,death or recovered.
\item \strong{province}: Name of province/state, for countries.
\item \strong{Month}: Each month variable.
\item \strong{Country}: Name of country/region.
\item \strong{long}: Longitude of center of geographic region.
\item \strong{lat}:Latitude of center of geographic region.
\item \strong{Cases}:Number of cases on given month.
}
}
\source{
\href{https://github.com/RamiKrispin/coronavirus}{Coronavirus Github}
}
\usage{
covid19_count
}
\description{
This data set displays the monthly coronavirus cases in each country with each region in each country
}
\keyword{datasets}
