% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19_region.R
\docType{data}
\name{covid19_region}
\alias{covid19_region}
\title{covid19_region}
\format{
a tibble with 51 observations and 6 variables
\itemize{
\item \strong{Type}: COVID19 confirmed cases.
\item \strong{province}: Name of province/state, for countries.
\item \strong{Country}: Name of country/region.
\item \strong{long}: Longitude of center of geographic region.
\item \strong{lat}:Latitude of center of geographic region.
\item \strong{Cases}:Number of confirmed cases on given region.
}
}
\source{
\href{https://github.com/RamiKrispin/coronavirus}{Coronavirus Github}
}
\usage{
covid19_region
}
\description{
This data set shows the cumulative COVID19 confirmed cases for each region in three countries.
}
\keyword{datasets}
