% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_option.R
\name{input_option}
\alias{input_option}
\title{Create input options for shiny app}
\usage{
input_option(input, data)
}
\arguments{
\item{input}{The variable can be chosen for the input selection via \code{"country"}, \code{"province"}, or \code{"summary"}.}

\item{data}{The data set for the input to be used interactively via \code{covid19_count} or \code{covid19_region}.}
}
\value{
Using the input options in shiny app selection could be visualized the content more interactive.
}
\description{
Create the shiny input options for the different variables which are contained in the data sets.
}
