% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Running Shiny App of COVID19 tracker}
\usage{
launch_app()
}
\description{
This function will launch shiny app where users can explore the case situation of Covid-19 for each region in three countries.
}
\examples{
\dontrun{
launch_app()
}


}
