% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\title{Summary_table}
\usage{
summary_table(country)
}
\arguments{
\item{Country}{The country data to be formatted in table such as Australia, China and United Kingdom.}
}
\value{
A formatted table of the data set will be displayed
}
\description{
Creates a styled and formatted table to represent the summary case situation of COVID19 COVID19 in each country.
}
